all: update review comments + show type for BeginTransaction + remove prints
